refactor: change structure hierarchy of editorshell and its dependencies
