+++ src/routes/Boards/Editor/BoardEditor/BoardEditor.scss
