+++ src/routes/Boards/Editor/EditorShell.scss
