+++ src/routes/Boards/Editor/TemplateEditor/index.ts
@@ -1,0 +1,1 @@
+export * from "./TemplateEditor";
