+++ src/routes/Boards/Editor/index.ts
@@ -1,0 +1,3 @@
+export * from "./EditorShell";
+export * from "./BoardEditor";
+export * from "./TemplateEditor";
